+++ packages/react/tsup.config.ts
@@ -1,13 +1,13 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-	entry: ["src/index.ts"],
-	format: ["esm", "cjs"],
-	dts: true,
-	splitting: false,
-	sourcemap: true,
-	clean: true,
-	minify: false,
-	target: "es2022",
-	external: ["react", "react-dom", "@hypermemory/core", "@hypermemory/visualizer-core"],
+  entry: ["src/index.ts"],
+  format: ["esm", "cjs"],
+  dts: true,
+  splitting: false,
+  sourcemap: true,
+  clean: true,
+  minify: false,
+  target: "es2022",
+  external: ["react", "react-dom", "@runstack-ai/hypermemory-core", "@runstack-ai/hypermemory-visualizer-core"],
 });